requirements.txt
pandas
selenium
pyautogui
sklearn
asyncio
pyppeteer2
beautifulsoup4